backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
snowflake-connector-python==3.12.2
requests==2.32.3
schedule==1.2.1
python-dotenv==1.0.1
pydantic==2.9.2
pytz==2024.2

# Dev tools (optional)
black==24.8.0
flake8==7.1.1
pytest==8.3.2
httpx==0.27.2
